+++ src/networks/nodes/avax-c-mainnet.js
@@ -1,0 +1,10 @@
+import { AVAX } from '../types';
+export default {
+  type: AVAX,
+  service: 'avax.network',
+  url: 'https://api.avax.network/ext/bc/C/rpc',
+  port: 443,
+  auth: false,
+  username: '',
+  password: ''
+};