+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+### Release v5.7.23
+
+### Bug
+* Remove tokens with no balance so users can manually add and load tokens [#2892](https://github.com/MyEtherWallet/MyEtherWallet/pull/2892)
+* Fix checking gas when swap from is not on ethereum [#2879](https://github.com/MyEtherWallet/MyEtherWallet/pull/2879)
+
+### Devop
+
+* consolidate node PRs [#2883](https://github.com/MyEtherWallet/MyEtherWallet/pull/2883)
+* Address no decimals for NEO [#2783](https://github.com/MyEtherWallet/MyEtherWallet/pull/2783)
+* change the way we are checking lend balance[#2893](https://github.com/MyEtherWallet/MyEtherWallet/pull/2893)
 ### Release v5.7.22
 
 ### Bug